紀錄可刪除與改名、清除前確認，任務欄加最近任務選單
- 紀錄每一筆按右鍵可以修改任務名稱或刪除
- 「清除紀錄」改成先確認，避免一按就全部刪掉
- 紀錄上限從 200 筆提高到一萬筆，原本約一個月前的紀錄會被悄悄刪掉
- 任務欄右邊加一個選單，列出最近用過的 6 個任務，點一下就填好
- 任務名稱統一去掉前後空白，空的叫「未命名」
- 修正新版 Swift 編譯器對 CGFloat 乘 cos／sin 回報的模稜兩可錯誤，
  三角函數的結果一律明確轉成 CGFloat

+++ Sources/Theme.swift
@@ -77,8 +77,8 @@ struct Ticks: Shape {
 
             // 刻度從外緣往內長，中心落在外緣與內端的中點
             let mid = radius - height / 2
-            let position = CGPoint(x: center.x + sin(angle) * mid,
-                                   y: center.y - cos(angle) * mid)
+            let position = CGPoint(x: center.x + CGFloat(sin(angle)) * mid,
+                                   y: center.y - CGFloat(cos(angle)) * mid)
 
             let tick = Path(roundedRect: CGRect(x: -width / 2, y: -height / 2,
                                                 width: width, height: height),
